requirements.txt
fastapi==0.91.0
routes==2.5.1
uvicorn==0.20.0
aiohttp==3.8.1
pydash==5.1.0
pydantic
Pillow
pilmoji-fixed
deep_translator